Converted AABB to int.

+++ MoonBase/Framework/Data/AABB.h
@@ -6,15 +6,15 @@ class AABB
 public:
 
 	AABB();
-	AABB(float x, float y, float w, float h);	//	Defines an axis-aligned bounding box with top-left (x,y) and dimensions (w,h)
+	AABB(int x, int y, int w, int h);	//	Defines an axis-aligned bounding box with top-left (x,y) and dimensions (w,h)
 	~AABB();
 
-	float Right();
-	float Bottom();
+	int Right();
+	int Bottom();
 
-	float _X, _Y, _Width, _Height;
+	int _X, _Y, _Width, _Height;
 
-	bool Contains(float x, float y);
+	bool Contains(int x, int y);
 	bool Contains(AABB box);
 	bool Intersects(AABB box);
 };
